If $QMAILQUEUE is set, invoke it instead of qmail-queue
This could be used to add a program into the qmail-smtpd->qmail-queue
pipeline that could do filtering, rewrite broken headers, etc.

Allows you to do things like virus scanning and RBL checking before the
message ever goes into the queue. If this other program returns a value
other then zero, the SMTP server treats it as a hard or soft error
(depending on the return value) and refuses to accept the message.

[Description merged from netqmail, original patch, and jms1]