PATCH maxrcpt.patch
The qmail-1.03-maxrcpt.patch patch allows you to set a limit on how many
recipients are specified for any one email message. RFC 2821 section
4.5.3.1 says that an MTA MUST allow at least 100 recipients for each
message, but MAY elect not to allow any more than that. I choose not to,
since this is one of the favourite tricks of the spammer.

To set the limit, create a /var/qmail/control/maxrcpt file which contains
the maximum number of recipients you wish to allow. If you don't want any
limit, you can either delete this file, or use 0 as the value.

With the "6c6" or later version of the combined patch, you can override
this value with the MAXRCPT environment variable. Also, a minimum limit of
100 recipients is enforced within the code- if you try to set it to a lower
number, it will use 100.

[desc by jms]