Recognize 0.0.0.0 as a local address
Fixes a difference between how Linux interprets the IP address 0.0.0.0
and how the *BSD systems handle it.

According to RFC 1122, the IP address 0.0.0.0 should always be treated as
an address for "this host, this network".

Part of qmail's loop-detection logic is determining whether or not a given
IP address "is" the current machine. This patch "teaches" qmail that
0.0.0.0 is always the local machine.

[Description by John M. Simpson <jms1@jms1.net>]