qmail-verify
This patchset adds a feature to qmail-smtpd's behaviour, enabled by an
environment variable: Each recipient specified in the SMTP session (RCPT
command) may be verified to see if it is a valid address on the system. A
new daemon, qmail-verify is used by qmail-smtpd to determine if addresses
are valid; qmail-smtpd communicates with qmail-verify using UDP.

http://free.acrconsulting.co.uk/sw/qv/qmail-verify_v1.32.patch
http://free.acrconsulting.co.uk/email/qmail-verify.html

Security:

Here are some notes relating to security with this patchset, these notes
are unlikely to be exhaustive:

It is conceivable that an attacker or spammer could abuse a system patched
using this patchset to mount dictionary attacks. To vary how a
qmail-verify patched system responds, you should make sure you understand
the difference between immediate and DEFERred recipient verification and
set your system appropriately. Immediate verification is likely to be more
attractive to an attacker/spammer.

On the other hand, with this patchset you have a greater degree of
certainty that messages in the queue for local users are destined for real
mailboxes.

This patchset deliberately splits the functionality of the original
realrcptto patch such that the address verification part runs as a
separate daemon, qmail-verify (running as root), with qmail-smtpd still
able to run as user qmaild. This also means that the changes to
qmail-smtpd are minimal.

There is no attempt made to check that the replies to the queries to
qmail-verify are genuine: It is conceivable that an attacker might wish to
generate spoof UDP replies to tell qmail-smtpd that an address is valid.
For this to occur, the attacker would either need to be a 'local' user to
generate such 'replies' on the loopback interface (default configuration),
or qmail-verify would need to be running on an external interface. In the
latter case it should be possible to block the qmail-verify port at a
firewall.

All that said however, such an attack is of limited use, since qmail will
still bounce the message at delivery time for an invalid recipient.

[Description by Gerrit Pape <pape@smarden.org>]